Move non-public classes to "internal" package
