+++ bundles/org.corpus_tools.hexatomic.edit.graph/src/main/java/org/corpus_tools/hexatomic/graph/GraphEditor.java
@@ -42,6 +42,11 @@ import javax.inject.Inject;
 import org.corpus_tools.hexatomic.console.AtomicalConsole;
 import org.corpus_tools.hexatomic.core.ProjectManager;
 import org.corpus_tools.hexatomic.core.errors.ErrorService;
+import org.corpus_tools.hexatomic.graph.internal.GraphDragMoveAdapter;
+import org.corpus_tools.hexatomic.graph.internal.RootTraverser;
+import org.corpus_tools.hexatomic.graph.internal.SaltGraphContentProvider;
+import org.corpus_tools.hexatomic.graph.internal.SaltGraphLayout;
+import org.corpus_tools.hexatomic.graph.internal.SaltGraphStyler;
 import org.corpus_tools.salt.SALT_TYPE;
 import org.corpus_tools.salt.common.SDocument;
 import org.corpus_tools.salt.common.SDocumentGraph;
@@ -59,7 +64,6 @@ import org.corpus_tools.salt.extensions.notification.Listener;
 import org.corpus_tools.salt.graph.GRAPH_ATTRIBUTES;
 import org.corpus_tools.salt.graph.IdentifiableElement;
 import org.corpus_tools.salt.util.DataSourceSequence;
-import org.eclipse.draw2d.MouseMotionListener;
 import org.eclipse.draw2d.Viewport;
 import org.eclipse.draw2d.geometry.Dimension;
 import org.eclipse.draw2d.geometry.Point;
