+++ bundles/org.corpus_tools.hexatomic.edit.graph/src/main/java/org/corpus_tools/hexatomic/graph/internal/RootTraverser.java
@@ -1,4 +1,4 @@
-package org.corpus_tools.hexatomic.graph;
+package org.corpus_tools.hexatomic.graph.internal;
 
 import java.util.Arrays;
