+++ bundles/org.corpus_tools.hexatomic.edit.graph/src/main/java/org/corpus_tools/hexatomic/graph/internal/SaltGraphContentProvider.java
@@ -19,7 +19,7 @@
  * #L%
  */
 
-package org.corpus_tools.hexatomic.graph;
+package org.corpus_tools.hexatomic.graph.internal;
 
 import java.util.List;
 import org.corpus_tools.salt.common.SDocumentGraph;
