+++ bundles/org.corpus_tools.hexatomic.edit.graph/src/main/java/org/corpus_tools/hexatomic/graph/internal/SaltGraphLayout.java
@@ -18,7 +18,7 @@
  * #L%
  */
 
-package org.corpus_tools.hexatomic.graph;
+package org.corpus_tools.hexatomic.graph.internal;
 
 import com.google.common.collect.BiMap;
 import com.google.common.collect.ComparisonChain;
